+++ src/Testing.Commons.Tests/Configuration/ConfigurationTester.net.cs
