+++ src/Testing.Commons.Tests/Configuration/Support/ExpirationElement.net.cs
