+++ src/Testing.Commons.Tests/Configuration/Support/ICachingConfiguration.net.cs
