+++ src/Testing.Commons.Tests/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Testing.Commons.Tests.core
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
