+++ src/Testing.Commons.Tests/Serialization/RoundtripBinarySerializerTester.net.cs
