+++ src/Testing.Commons.Tests/Serialization/RoundtripJsonSerializerTester.net.cs
