+++ src/Testing.Commons.Tests/Serialization/Subjects/Serializable.cs
@@ -2,8 +2,7 @@
 
 namespace Testing.Commons.Tests.Serialization.Subjects
 {
-	[Serializable]
-	public class Serializable
+	public partial class Serializable
 	{
 		public string S { get; set; }
 		public decimal D { get; set; }
