+++ src/Testing.Commons.Tests/Serialization/Subjects/Serializable.net.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Testing.Commons.Tests.Serialization.Subjects
+{
+	[Serializable]
+	public partial class Serializable
+	{
+	}
+}
